requirements.txt
# PIP requirements
websocket-client==0.37.0
requests==2.11.1
